refactor: changes queue folder name
